+++ src/ng2-table/filter-input.component.js
@@ -41,7 +41,7 @@ FilterInputComponent = __decorate([
     core_1.Component({
         selector: 'filter-input-cmp',
         styles: ["\n    .form-control {\n      max-width: 200px;\n      font-weight: 400;\n    }\n  "],
-        template: "\n    <input type=\"text\"\n      [value]=\"filterValue || ''\"\n      class=\"form-control\"\n      (keyup)=\"valueChange($event)\"\n      placeholder=\"{{placeholder}}\">\n  "
+        template: "\n    <input type=\"text\"\n      [value]=\"filterValue || ''\"\n      class=\"form-control\"\n      (keyup)=\"valueChange($event)\"\n      placeholder=\"{{placeholder || ''}}\">\n  "
     }),
     __metadata("design:paramtypes", [])
 ], FilterInputComponent);