+++ src/example/app.component.js
@@ -5,9 +5,6 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
-};
 var core_1 = require("@angular/core");
 var example_table_data_1 = require("./example-table.data");
 var AppComponent = (function () {
@@ -32,8 +29,7 @@ AppComponent = __decorate([
     core_1.Component({
         selector: 'app',
         template: "\n    <example-table-component>\n    </example-table-component>\n\n    <br><br>\n    <button class=\"btn btn-secondary btn-sm\" (click)=\"showMinimalTableFn()\">\n      Show minimal config example\n    </button>\n\n    <ng2-table\n      *ngIf=\"showMinimalTable\"\n      [tableData]=\"exampleData\"\n      [tableConfig]=\"minimalTableConfig\">\n    </ng2-table>\n  "
-    }),
-    __metadata("design:paramtypes", [])
+    })
 ], AppComponent);
 exports.AppComponent = AppComponent;
 //# sourceMappingURL=app.component.js.map