+++ src/ng2-table/ng2-table.module.js
@@ -5,9 +5,6 @@ var __decorate = (this && this.__decorate) || function (decorators, target, key,
     else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;
     return c > 3 && r && Object.defineProperty(target, key, r), r;
 };
-var __metadata = (this && this.__metadata) || function (k, v) {
-    if (typeof Reflect === "object" && typeof Reflect.metadata === "function") return Reflect.metadata(k, v);
-};
 var core_1 = require("@angular/core");
 var platform_browser_1 = require("@angular/platform-browser");
 var ng2_table_component_1 = require("./ng2-table.component");
@@ -27,8 +24,7 @@ Ng2TableModule = __decorate([
             filter_pipe_1.FilterPipe
         ],
         exports: [ng2_table_component_1.Ng2TableComponent]
-    }),
-    __metadata("design:paramtypes", [])
+    })
 ], Ng2TableModule);
 exports.Ng2TableModule = Ng2TableModule;
 //# sourceMappingURL=ng2-table.module.js.map